Quesito 01 feito

+++ Trabalho 03/avl.h
@@ -1,5 +1,5 @@
 typedef struct noAVL {
-    int d;
+    float d;
     int indice;
     int altura;
     struct noAVL *esq;
